% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{guessVideoFormat}
\alias{guessVideoFormat}
\title{Guess Video Format Type}
\usage{
guessVideoFormat(files)
}
\arguments{
\item{files}{A vector of URL paths (relative or absolute) to videos}
}
\value{
A vector the same length as \code{files} of the video types.
}
\description{
If no type is provided when generating a video.js video, then the format needs to
be guessed. Included in the package is a dataset of the default type of each
video. This will give the default type of each file provided.
}
\examples{
guessVideoFormat("video.mp4")

}
