% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-side.R
\name{video-server}
\alias{video-server}
\alias{playVideo}
\alias{pauseVideo}
\alias{stopVideo}
\alias{seekVideo}
\alias{changeVideo}
\title{Update video.js Server-Side}
\usage{
playVideo(id, session = getDefaultReactiveDomain())

pauseVideo(id, session = getDefaultReactiveDomain())

stopVideo(id, session = getDefaultReactiveDomain())

seekVideo(id, seek, session = getDefaultReactiveDomain())

changeVideo(id, files, format = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of the \code{video} to update}

\item{session}{Shiny session}

\item{seek}{Time (in seconds) to set the position of the track}

\item{files}{A vector of file paths or URLs pointing}

\item{format}{An optional list of formats of \code{video}}
}
\value{
Updates the the state of the specified \code{video} in the shiny application.
}
\description{
Change the state of the video player from the server.

\code{playVideo}, \code{pauseVideo} and \code{stopVideo} will all be applied to the current track.

\code{changeTrack} will update the track to the file specified. This file must be included when the player
is initialised, otherwise it won't change the track.

\code{addTrack} will add a new track to the specified player.
}
\examples{
if (interactive()) {}

}
