% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{video_formats}
\alias{video_formats}
\title{Video Formats}
\format{
A data frame with 88 rows and 2 variables:
\describe{
\item{extension}{File extension string}
\item{type}{Type attribute to give to a <video> HTML tag}
}
}
\usage{
video_formats
}
\description{
A datset containing the format type of all potential video file extensions. Used
as the "type" attribute for HTML videos.
}
\keyword{datasets}
