% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-side.R
\name{video-server}
\alias{video-server}
\alias{playHowl}
\alias{pauseHowl}
\alias{stopHowl}
\alias{seekVideo}
\title{Update video.js Server-Side}
\usage{
playHowl(session, id)

pauseHowl(session, id)

stopHowl(session, id)

seekVideo(session, id, seek)
}
\arguments{
\item{session}{Shiny session}

\item{id}{ID of the \code{video} to update}

\item{seek}{Time (in seconds) to set the position of the track}
}
\value{
Updates the the state of the specified \code{video} in the shiny application.
}
\description{
Change the state of the video player from the server.

\code{playVideo}, \code{pauseVideo} and \code{stopVideo} will all be applied to the current track.

\code{changeTrack} will update the track to the file specified. This file must be included when the player
is initialised, otherwise it won't change the track.

\code{addTrack} will add a new track to the specified player.
}
\examples{
if (interactive()) {}

}
