% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{video}
\alias{video}
\title{Video Player}
\usage{
video(
  files,
  format = NULL,
  options = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{files}{A vector of file paths or URLs pointing}

\item{format}{An optional list of formats of \code{video}}

\item{options}{A named list of options to apply to the video. List of available options
available in Details}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{HTML id tag to be given to the video player element}
}
\description{
A video player that can be embedded in HTML pages.
}
