% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoWidget.R
\name{video}
\alias{video}
\title{Creating a Video widget}
\usage{
video(video = NULL, elementId = NULL, width = NULL, height = NULL,
  options = NULL)
}
\arguments{
\item{elementId}{ID for the element in the DOM}

\item{width}{Width of the widget}

\item{height}{height of the widget}

\item{img}{Vector of character file names or list of \code{nifti} images}
}
\description{
Wraps a widget for the Video JavaScript library
}
\examples{


}
